LoginForm/app/requirements.txt
Django
gunicorn
django-heroku
whitenoise
dj-database-url
djangorestframework
djangorestframework-simplejwt
pytz
pyjwt
